requirements.txt
twscrape
asyncio
google-api-python-client
yt-dlp
python-dotenv
tenacity
youtube-transcript-api
linkedin-api
